Create concrete Repository implementation as gateway to Blog data
I'd like to encapsulate data storage behind an interface that allows me
to easily test the application and also gives me flexibility to write
adapters for a wide range of deployment scenarios.

Create a Repository class to implement methods corresponding to specific
data access patterns of the application, such as getting all_sites.