Install webpack dev server for live reloading
I'd like to be able to develop with a quick feedback loop, especially
for visual changes, such as changes to stylesheets.

Setup webpack's dev server for live reloading of a page while editing
styles:

    npm install --save-dev webpack-dev-server

Add "npm run start" command to start it.  Configure it to also reload
the development browser when HTML files generated by Towhee::Prerender
(not by webpack) change in the dist folder with watchContentBase.