Use mini-css-extract-plugin to extract CSS from JS for production
In development we need to embed CSS in a JS file to support live reload,
the super-slick feature that causes HTML, CSS, and JS to be replaced in
a running browser without a page refresh whenever a file on disk
changes.

On the other hand, in production, we want to load CSS directly from a
CSS file.  So we need a way to build a separate file for production
only.

Use mini-css-extract-plugin to extact the CSS that would have gone into
the JS file into its own CSS file.