Link to webpack's JS and CSS bundles from HomeView
HomeView is still linking to the style.css file that is
no longer being built.  Update it to load main.js and main.css, as built
by webpack, instead of style.css, so that it will load styles whether
in development or production.  (In a clean development asset build, it
will fail to load main.css, but that won't be very disruptive.  We might
stop linking to it from HomeView in development mode at some later
point, if we develop a more complete notion of development mode there.)