Extract multi-table repository's SQL details to an adapter
The MultiTableInheritance::Repository is getting lengthy, and it's hard
to write specs in terms of sql queries sent to the active record
adapter.

Extract an additional adapter layer consisting of four Repository
methods that invoke the ActiveRecord database adapter.  This will lay
the ground work for addition an in-memory adapter for easier testing.