Fix bug when id != entity_id for a descendant row
While converting the repository_spec to use the memory_adapter instead
of an active_record_adapter with mocked connection_adapter, I
encountered this error:

  1) Towhee::MultiTableInheritance::Repository multiple records loads a record
     Failure/Error:
       row = row.merge(rows_by_type.
                       fetch(type).
                       fetch(row.fetch("id")))

     KeyError:
       key not found: 5
     # lib/towhee/multi_table_inheritance/repository.rb:48:in `fetch'
     # lib/towhee/multi_table_inheritance/repository.rb:79:in `walk_lineage'
     # lib/towhee/multi_table_inheritance/repository.rb:44
     # lib/towhee/multi_table_inheritance/repository.rb:41:in `map'
     # lib/towhee/multi_table_inheritance/repository.rb:41:in `find_all'
     # spec/multi_table_inheritance/repository_spec.rb:61

The memory adapter differs from the mocked-up AR adapter in that it
allocates IDs from a global sequence rather than a per-table sequence.
As a result, this example was not able to coincidentally pass because
this row's entity_id happened to equal its id.

Adjust find_all to use the correct ID field -- "id" for the "entities"
table, and "entity_id" for other tables.