Update repository_spec to eliminate verbose SQL mocks
The repository_spec is difficult to manage because it contains so many
SQL strings.  These are a bit fragile and perhaps too low-level.  But
most of all, they are verbose.

Replace those mocked SQL queries with data seeded in the MemoryAdapter.
Allow Numeric values because they're needed to store entity_ids.