Add ability to query by any attribute via inner join
We have the ability to lookup entities by ID, but not by other
attributes.

Add the ability to lookup entities by any attribute.  Use the
cornerstone pattern of joining across the tables for each type in the
path from the desired type to the root type.

We are faced with a design tradeoff of whether to return entities as
their concrete type (when asking for Site, return a blog as a Blog), or
whether to return the entities as the requested type, which may be a
parent type of the concrete type of the entity.  Unfortunately
returning the entities in their concrete type would require either
joining all known tables to include all needed attributes or refetching
the entities by ID once their types are known.  The first approach may
involve over-fetching, and the second requires additional queries.  On
the other hand, returning types as the more abstract type allows us to
join tables for a single path from that type to the root type, which is
likely far fewer tables.  But it means that the entity may have
incorrect behavior if its methods are expected to be overridden by the
more concrete class.  Choose the latter tradeoff (returning the abstract
type) for now, leaving open the possibility for adding support for the
other way later.