Render Blog::PostViews in Blog::App#render
Add Post pages to what gets rendered by the Blog::App#render method.

Somehow we now need a key on the HomeView, perhaps because #uniq doesn't
need to call #key when there was only one item?  But we didn't need it
previously.

Hard-code the association from posts to sites to always map to the first
site for now, because the repository doesn't know which posts are
associated with which posts yet.  But so far we only have one site, so
we can take a shortcut.  Someday this will choose a site based on the
post.  And the rest of the app needn't change when we remove the
shortcut.