Make view constructor convention uniform by using keys
The HomeView constructor doesn't use a key for its primary model, which
is inconsistent with the PostView, and it leaves the reader wondering
what the "model" means to the view when passed in the ViewEnumerator
(unless you realize you can glean it from the "when" condition).

Change the HomeView site parameter from a required positional parameter
to a required kwargs parameter.