Store association between sites and posts
So far, we've been assuming that all posts belong to the first site in
the Blog::Repository.  However, this is incomplete functionality; we
want to be able to associate blogs to multiple sites.

Convey the associations to the Repository by passing to its constructor
a Hash of sites to posts.  Adjust the #post_site method to use this new
information, and upgrade call sites of the constructor.

We could have used IDs to convey the association.  While it might be
necessary for a case where posts are stored in the database, it is not
necessary for this sample in-memory repository, and it polutes the Plain
Old Ruby Objects (PORO) of the domain model with a persistence-related
ID accessor.