Make posts required for Blog::HomeView
It's nice to be reminded when forgetting to pass data expected by a
view.  Requiring keys to be passed rather than making them optional can
help with this.

Require the :posts key to the HomeView constructor instead of allowing
it to default to an empty list.  While this is a sensible degenerate
default, I can't think of a case where it makes sense for the caller to
invoke the view with not even an empty array.