Extract parts of HomeView to Blog::Layout
Just as we've extracted parts of PostView to Layout, there are parts of
HomeView that we'd like to reuse across pages in the Layout.  These
parts build on what is already present in the Layout for PostView, as
HomeView's page structure was a bit more evolved in that it had a header
and Twitter Bootstrap-style column and container classes.

Move column and container markup to the Layout and move the header to
the Layout.  Add a new parameter -- :site -- to be used in the header.
Although the header used to only be present on the HomeView, let it be
present for all contexts (including PostView).