Test Blog::Layout structure using REXML to parse the HTML
So far, I've gotten by with just Regexps do test view output.  This
feels clumsy but doesn't require any external dependencies.  But with
the Layout, "contains" relationships become more important.

Add support for testing containment of text within elements.  For any
text that is not a direct child of that element, this is hard to do with
Regexps, so use REXML instead.

Use REXML instead of nokogiri, which may be better suited to the job, to
avoid an external dependency of this library.  If this becomes a common
need, maybe we can give the user a choice of driver someday.