Make Blog::Layout easy to override by derivative blogs
I'd like to make my own personal blog, but of course it has to look
different from the default blog.  So I'd like a way to easily override
the Layout.

Create a factory method for the layout class and the view_enumerator so
they can both be easily overridden by an extending application.